+++ src/main/java/org/oxff/repeater/privilege/report/ReportContainerReader.java
@@ -9,7 +9,6 @@ import java.awt.Frame;
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.zip.CRC32;
 import java.util.zip.Inflater;