+++ src/main/java/org/oxff/repeater/privilege/report/ReportContainerWriter.java
@@ -5,7 +5,6 @@ import org.oxff.repeater.logging.LogManager;
 import java.awt.Component;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.ByteBuffer;